worked on missions filter
