--- src/routes/MyProfile.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function MyProfile() {
-  return (
-    <div>MyProfile</div>
-  );
-}
-
-export default MyProfile;
